+++ cucumber-core/src/main/java/io/cucumber/core/plugin/TeamCityPlugin.java
@@ -45,6 +45,7 @@ import java.util.stream.Collectors;
 
 import static io.cucumber.core.exception.ExceptionUtils.printStackTrace;
 import static java.util.Collections.emptyList;
+import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.joining;
 
 /**
@@ -77,6 +78,9 @@ public class TeamCityPlugin implements EventListener {
             + "[testFinished timestamp = '%s' duration = '%s' name = '%s']";
     private static final String TEMPLATE_TEST_FAILED = TEAMCITY_PREFIX
             + "[testFailed timestamp = '%s' duration = '%s' message = '%s' details = '%s' name = '%s']";
+
+    private static final String TEMPLATE_TEST_COMPARISON_FAILED = TEAMCITY_PREFIX
+            + "[testFailed timestamp = '%s' duration = '%s' message = '%s' details = '%s' expected = '%s' actual = '%s' name = '%s']";
     private static final String TEMPLATE_TEST_IGNORED = TEAMCITY_PREFIX
             + "[testIgnored timestamp = '%s' duration = '%s' message = '%s' name = '%s']";
 
@@ -100,6 +104,24 @@ public class TeamCityPlugin implements EventListener {
 
     private static final Pattern ANNOTATION_GLUE_CODE_LOCATION_PATTERN = Pattern.compile("^(.*)\\.(.*)\\([^:]*\\)");
     private static final Pattern LAMBDA_GLUE_CODE_LOCATION_PATTERN = Pattern.compile("^(.*)\\.(.*)\\(.*:.*\\)");
+
+    private static final Pattern[] COMPARE_PATTERNS = new Pattern[] {
+            // Hamcrest 2 MatcherAssert.assertThat
+            Pattern.compile("expected: (.*)(?:\r\n|\r|\n) {5}but: was (.*)$",
+                Pattern.DOTALL | Pattern.CASE_INSENSITIVE),
+            // AssertJ 3 ShouldBeEqual.smartErrorMessage
+            Pattern.compile("expected: (.*)(?:\r\n|\r|\n) but was: (.*)$",
+                Pattern.DOTALL | Pattern.CASE_INSENSITIVE),
+            // JUnit 5 AssertionFailureBuilder
+            Pattern.compile("expected: <(.*)> but was: <(.*)>$",
+                Pattern.DOTALL | Pattern.CASE_INSENSITIVE),
+            // JUnit 4 Assert.assertEquals
+            Pattern.compile("expected:\\s?<(.*)> but was:\\s?<(.*)>$",
+                Pattern.DOTALL | Pattern.CASE_INSENSITIVE),
+            // TestNG 7 Assert.assertEquals
+            Pattern.compile("expected \\[(.*)] but found \\[(.*)]\n$",
+                Pattern.DOTALL | Pattern.CASE_INSENSITIVE),
+    };
 
     private final PrintStream out;
     private final List<SnippetsSuggestedEvent> suggestions = new ArrayList<>();
@@ -281,7 +303,18 @@ public class TeamCityPlugin implements EventListener {
             case AMBIGUOUS:
             case FAILED: {
                 String details = printStackTrace(error);
-                print(TEMPLATE_TEST_FAILED, timeStamp, duration, "Step failed", details, name);
+                String message = error.getMessage();
+                if (message == null) {
+                    print(TEMPLATE_TEST_FAILED, timeStamp, duration, "Step failed", details, name);
+                    break;
+                }
+                ComparisonFailure comparisonFailure = ComparisonFailure.parse(message.trim());
+                if (comparisonFailure == null) {
+                    print(TEMPLATE_TEST_FAILED, timeStamp, duration, "Step failed", details, name);
+                    break;
+                }
+                print(TEMPLATE_TEST_COMPARISON_FAILED, timeStamp, duration, "Step failed", details,
+                    comparisonFailure.getExpected(), comparisonFailure.getActual(), name);
                 break;
             }
             default:
@@ -420,4 +453,43 @@ public class TeamCityPlugin implements EventListener {
                 .replace("]", "|]");
     }
 
+    private static class ComparisonFailure {
+
+        static ComparisonFailure parse(String message) {
+            for (Pattern pattern : COMPARE_PATTERNS) {
+                ComparisonFailure result = parse(message, pattern);
+                if (result != null) {
+                    return result;
+                }
+            }
+            return null;
+        }
+
+        static ComparisonFailure parse(String message, Pattern pattern) {
+            final Matcher matcher = pattern.matcher(message);
+            if (!matcher.find()) {
+                return null;
+            }
+            String expected = matcher.group(1);
+            String actual = matcher.group(2);
+            return new ComparisonFailure(expected, actual);
+        }
+
+        private final String expected;
+
+        private final String actual;
+
+        ComparisonFailure(String expected, String actual) {
+            this.expected = requireNonNull(expected);
+            this.actual = requireNonNull(actual);
+        }
+
+        public String getExpected() {
+            return expected;
+        }
+
+        public String getActual() {
+            return actual;
+        }
+    }
 }